Some work on front end, still buggy ws

+++ shared/src/lib.rs
@@ -1,3 +1,4 @@
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq)]
 pub enum ClientMessage {
     LinesDestroyed (
         // u16 since it's boolean anyway, we can use an int as an array of bits, don't need sign nor more than 16 wide
@@ -7,7 +8,8 @@ pub enum ClientMessage {
     GameOver,
 }
 
+#[derive(Debug, serde::Serialize, serde::Deserialize, PartialEq, Eq)]
 pub enum ServerMessage {
-    Broadcast { user_id: i32, msg: ClientMessage },
+    Broadcast { user_id: u128, msg: ClientMessage },
     LeaderBoardUpdate(),
 }
